populate predicates
